Avoid deleting snapshot if the input name is invalid, i.e. name that does not starts with "snapshot-". This prevents the possibility of deleting the root filesystem "/" using delete-snapshot action.

+++ src/charm.py
@@ -150,6 +150,14 @@ class AptMirrorCharm(CharmBase):
 
     def _on_delete_snapshot_action(self, event):
         snapshot = event.params["name"]
+        if not snapshot.startswith("snapshot-"):
+            event.set_results(
+                {
+                    "ReturnCode": 1,
+                    "Stderr": "Invalid snapshot name: {}".format(snapshot),
+                }
+            )
+            return
         logger.info("Delete snapshot {}".format(snapshot))
         shutil.rmtree("{}/{}".format(self._stored.config["base-path"], snapshot))
         self._update_status()
